third_party/WebKit/LayoutTests/inspector/elements/styles/inject-stylesheet-expected.txt

CONSOLE MESSAGE: line 9: iframe loaded
Tests that injected user stylesheets are reflected in the Styles pane.


Main frame style:
-webkit-background-clip: border-box;
-webkit-background-origin: padding-box;
-webkit-background-size: auto;
-webkit-border-image: url(data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAQAAAC1+jfqAAAAiElEQ…Ms+LS30CAhBN5nNxeT5hbJ1zwmji2k+aF6NENIPf/hs54f0sZFUVAMigAAAABJRU5ErkJggg==) 100% / 1 / 0px stretch;
background-attachment: scroll;
background-clip: border-box;
background-color: rgb(0, 0, 255);
background-image: none;
background-origin: padding-box;
background-size: auto;
color: rgb(255, 0, 0);
display: block;
[expanded] 
element.style  { ()

[expanded] 
media="screen" (inject-stylesheet.html)
#main  { (inject-stylesheet.html:5 -> inject-stylesheet.html:5:3)
  background: blue;
      background-image: initial;
      background-position-x: initial;
      background-position-y: initial;
      background-size: initial;
      background-repeat-x: initial;
      background-repeat-y: initial;
      background-attachment: initial;
      background-origin: initial;
      background-clip: initial;
      background-color: blue;

[expanded] 
#main  { (injected stylesheet)
  color: red;
  -webkit-border-image: url(data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAQAAAC1+jfqAAAAiElEQ…Ms+LS30CAhBN5nNxeT5hbJ1zwmji2k+aF6NENIPf/hs54f0sZFUVAMigAAAABJRU5ErkJggg==);

[expanded] 
div  { (user agent stylesheet)
  display: block;

iframe style:
-webkit-background-clip: border-box;
-webkit-background-origin: padding-box;
-webkit-background-size: auto;
background-attachment: scroll;
background-clip: border-box;
background-color: rgb(255, 0, 0);
background-image: none;
background-origin: padding-box;
background-size: auto;
display: block;
margin-bottom: 8px;
margin-left: 8px;
margin-right: 8px;
margin-top: 8px;
[expanded] 
element.style  { ()

[expanded] 
#iframebody  { (injected stylesheet)
  background: red;
      background-image: initial;
      background-position-x: initial;
      background-position-y: initial;
      background-size: initial;
      background-repeat-x: initial;
      background-repeat-y: initial;
      background-attachment: initial;
      background-origin: initial;
      background-clip: initial;
      background-color: red;

[expanded] 
media="screen" (inject-stylesheet-iframe-data.html)
body  { (inject-styleshe…me-data.html:4 -> inject-stylesheet-iframe-data.html:4:3)
  background: green;
    /-- overloaded --/   background-image: initial;
    /-- overloaded --/   background-position-x: initial;
    /-- overloaded --/   background-position-y: initial;
    /-- overloaded --/   background-size: initial;
    /-- overloaded --/   background-repeat-x: initial;
    /-- overloaded --/   background-repeat-y: initial;
    /-- overloaded --/   background-attachment: initial;
    /-- overloaded --/   background-origin: initial;
    /-- overloaded --/   background-clip: initial;
    /-- overloaded --/   background-color: green;

[expanded] 
body  { (user agent stylesheet)
  display: block;
  margin: 8px;
      margin-top: 8px;
      margin-right: 8px;
      margin-bottom: 8px;
      margin-left: 8px;
